% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BQMAResult-class}
\alias{BQMAResult-class}
\title{BQMAResult S4 class}
\description{
Stores the output of a BQMA model fit.
}
\section{Slots}{

\describe{
\item{\code{tau}}{Numeric vector of quantiles fitted.}

\item{\code{pip}}{Matrix (CpGs x tau) of posterior inclusion probabilities.}

\item{\code{beta_mean}}{Matrix (CpGs x tau) of posterior mean effects.}

\item{\code{beta_ci}}{Array (CpGs x tau x 2) of 95 percent credible intervals.}

\item{\code{n_iter}}{Integer. MCMC iterations after burnin.}

\item{\code{burnin}}{Integer. Burnin iterations discarded.}

\item{\code{call}}{The matched call.}

\item{\code{probe_class}}{Named character vector of probe annotations. Empty if
\code{annotate_results()} has not been called.}
}}

