% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getBeta}
\alias{getBeta}
\alias{getBeta,BQMAResult-method}
\title{Extract posterior mean beta coefficients}
\usage{
getBeta(x, tau = NULL, filter = NULL)

\S4method{getBeta}{BQMAResult}(x, tau = NULL, filter = NULL)
}
\arguments{
\item{x}{A \code{BQMAResult} object.}

\item{tau}{Numeric. A single quantile value. NULL returns all.}

\item{filter}{Character vector of \code{probe_class} values to keep.
NULL returns all probes.}
}
\value{
A matrix of posterior means (CpGs x tau).
}
\description{
Extract posterior mean beta coefficients
}
\section{Methods (by class)}{
\itemize{
\item \code{getBeta(BQMAResult)}: Method for BQMAResult

}}
