% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getCI}
\alias{getCI}
\alias{getCI,BQMAResult-method}
\title{Extract posterior credible intervals}
\usage{
getCI(x, tau = NULL, filter = NULL)

\S4method{getCI}{BQMAResult}(x, tau = NULL, filter = NULL)
}
\arguments{
\item{x}{A \code{BQMAResult} object.}

\item{tau}{Numeric. A single quantile value. NULL returns all.}

\item{filter}{Character vector of \code{probe_class} values to keep.
NULL returns all probes.}
}
\value{
A matrix with columns lower and upper (95 percent credible interval).
}
\description{
Extract posterior credible intervals
}
\section{Methods (by class)}{
\itemize{
\item \code{getCI(BQMAResult)}: Method for BQMAResult

}}
