% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,BQMAResult-method}
\alias{show,BQMAResult-method}
\title{Show method for BQMAResult}
\usage{
\S4method{show}{BQMAResult}(object)
}
\arguments{
\item{object}{A \code{BQMAResult} object.}
}
\description{
Show method for BQMAResult
}
