% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getBeta}
\alias{getBeta}
\alias{getBeta,BQMAResult-method}
\title{Extract posterior mean beta coefficients}
\usage{
\S4method{getBeta}{BQMAResult}(object, tau = NULL, filter = NULL)
}
\arguments{
\item{object}{A \code{BQMAResult} object.}

\item{tau}{Numeric. A single quantile value. NULL returns all.}

\item{filter}{Character vector of \code{probe_class} values to keep.
NULL returns all probes.}
}
\value{
A matrix of posterior means (CpGs x tau).
}
\description{
Method for \code{BQMAResult} of the \code{getBeta} generic imported
from \pkg{minfi}.
}
